requirements.txt
PySide6
PyOpenGL
numpy
pygame
